added mock tests
