+++ src/main/java/com/abbaspour/service/DecodeService.java
@@ -1,9 +1,13 @@
 package com.abbaspour.service;
 
 import com.abbaspour.model.BitMap;
+import com.abbaspour.utils.Decode;
 import org.springframework.stereotype.Service;
 
-
-public interface DecodeService {
-    BitMap decode(String hex);
+@Service
+public class DecodeService implements DecodeServiceInterface {
+    @Override
+    public BitMap decode(String hex) {
+        return Decode.generateBitMapFromPattern(hex);
+    }
 }
