--- src/main/java/com/abbaspour/service/DecodeServiceImpl.java
@@ -1,13 +1,0 @@
-package com.abbaspour.service;
-
-import com.abbaspour.model.BitMap;
-import com.abbaspour.utils.Decode;
-import org.springframework.stereotype.Service;
-
-@Service
-public class DecodeServiceImpl implements DecodeService{
-    @Override
-    public BitMap decode(String hex) {
-        return Decode.generateBitMapFromPattern(hex);
-    }
-}
