+++ src/main/java/com/abbaspour/service/DecodeServiceInterface.java
@@ -1,0 +1,8 @@
+package com.abbaspour.service;
+
+import com.abbaspour.model.BitMap;
+
+
+public interface DecodeServiceInterface {
+    BitMap decode(String hex);
+}
