+++ src/main/java/com/abbaspour/service/IntegerToWordServiceInterface.java
@@ -1,0 +1,6 @@
+package com.abbaspour.service;
+
+public interface IntegerToWordServiceInterface {
+    String toEnglishWordConverter(Long number);
+    String toFrenchWordConverter(Long number);
+}
